requirements.txt
discord.py
sqlalchemy
